Update default-bundled-gems test

+++ test/truffle/gems/default-bundled-gems.sh
@@ -7,6 +7,12 @@ cd "$truffle/gems/default-bundled-gems" || exit 1
 jt ruby gems2gemfile.rb
 
 cat Gemfile
+
+if [[ "$(jt ruby -S bundle config path)" =~ "Set for " ]]; then
+    echo "This test only works with no bundle path"
+    exit 2
+fi
+
 jt ruby -S bundle install --local
 
 echo 'Check if Gemfile is up to date'
